More Formatting and Basic Revisions
Removed parent lunch info, it is different this year due to the booster club.

Changed formatting on the silver page to be neater, still not happy with it, may bullet at a later date.

Modified meeting schedule: Now more fluid sentences and includes link to bvcaps on google maps. Removed link to silver page, unnecessary.